Fix the seven CodeQL alerts on the reorganized tree
    What changed: test/ingestion/html.test.mjs checks the extracted link
    list through a Set instead of Array.includes. The scaffold's worker.js
    and worker.artifact.js templates return a fixed "Internal server error"
    body on 5xx and log the error to the Worker console; 4xx/503 bodies
    carry error.message only when the thrown value is an Error. The legacy
    reference Worker drops its String(error) fallback and no longer appends
    the JSON parser's message to "Invalid JSON body". The Docusaurus
    plugin's slugifyName/trimSlashes strip leading/trailing characters with
    an index walk (trimChar) instead of ^x+|x+$ regexes. The quantization
    golden generator multiplies in double, and golden-quantization.json is
    regenerated from it. CHANGELOG records the template change.

    Why: CodeQL reports 13 alerts on this branch, 6 of them the workflow
    permissions closed in bee76e3. Of the rest, three are real: the
    templates echoed every unexpected error's message to callers, which
    can name bundled asset paths and load state; the trim regexes are
    quadratic on long runs of '-' or '/' in config-supplied site names and
    baseUrl. Two are analyzer misreads (Array.includes on a list of URLs;
    a float product accumulated into a double), fixed in the form the
    analyzer wants because the cost is nil.

    Expected behavior: Workers generated from these templates answer 5xx
    with { error: "Internal server error", code } and the message appears
    in wrangler tail; 4xx and 503 responses are unchanged. Slug and path
    trimming is byte-identical to the regex form (fuzzed 200k random
    strings, 0 mismatches). expectedExactDot in the golden moves at the
    8th-9th significant digit on 4 of 5 rows, inside the test's 0.02
    tolerance. Projects generated by earlier versions are untouched.

    Validation: npm test 1235 passed / 0 failed with every suite summary
    present (sketch 106, staged 13, row integrity 18, hardening 66,
    complete-profile 163, MCP 28, ingestion 75, tsc), EXIT=0. Scaffold e2e
    47/47 under wrangler against the edited templates. The old generator
    reproduces the committed golden byte-for-byte before the change; the
    conformance test passes on the regenerated file. The CodeQL result is
    read off the pull request's next analysis.

+++ benchmarks/beir/conformance/gen-golden-quantization.cpp
@@ -61,7 +61,10 @@ int main() {
         normalize_cosine(query.data(), queryNormed.data(), dim);
 
         double exactDot = 0, sumQuery = 0;
-        for (size_t d = 0; d < dim; d++) { exactDot += queryNormed[d] * normed[d]; sumQuery += queryNormed[d]; }
+        for (size_t d = 0; d < dim; d++) {
+            exactDot += static_cast<double>(queryNormed[d]) * static_cast<double>(normed[d]);
+            sumQuery += queryNormed[d];
+        }
 
         printf("  {\n");
         printf("    \"inputRow\": [");
